+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Use `reframe` in grouping summaries in `das_sight` for when none of a certain event are in the data (#8)
 
+* Ensure that `sum(GsSpx)` equals `GsSchoolx`, for all of Best, High and Low (#9)
+
 
 # swfscDAS 0.6.1